requirements.txt
python-telegram-bot
openai
yfinance
akshare
httpx
beautifulsoup4
matplotlib
markdown
weasyprint
python-dotenv
duckduckgo-search
asyncpg
google-genai
fastapi
uvicorn[standard]
PyJWT
bcrypt
python-multipart
fpdf2
